# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flash API",
    "version": "0.1.0",
    "description": "API de la plateforme de monnaie électronique Flash."
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
...
    "/v1/merchant/qr": {
      "get": {
        "summary": "QR statique du marchand (montant saisi par le client)",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/merchant/charges": {
      "post": {
        "summary": "Créer un QR dynamique (montant + référence + expiration)",
        "tags": [
          "merchants"
        ],
        "responses": {
          "201": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "amount_minor": {
                    "examples": [
                      25000
                    ],
                    "exclusiveMinimum": 0,
                    "title": "Amount Minor",
                    "type": "integer"
                  },
                  "reference": {
                    "examples": [
                      "Table 4"
                    ],
                    "maxLength": 80,
                    "minLength": 1,
                    "title": "Reference",
                    "type": "string"
                  },
                  "ttl_minutes": {
                    "default": 60,
                    "maximum": 1440,
                    "minimum": 1,
                    "title": "Ttl Minutes",
                    "type": "integer"
                  }
                },
                "required": [
                  "amount_minor",
                  "reference"
                ],
                "title": "CreateChargeRequest",
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/v1/merchant/kyb": {
      "post": {
        "summary": "Soumettre (ou re-soumettre) le dossier de vérification marchand",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/merchant/api-keys": {
      "post": {
        "summary": "Émettre une clé d'API marchande (le secret n'est montré qu'ici)",
        "tags": [
          "merchants"
        ],
        "responses": {
          "201": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "label": {
                    "default": "",
                    "examples": [
                      "Caisse principale"
                    ],
                    "maxLength": 60,
                    "title": "Label",
                    "type": "string"
                  }
                },
                "title": "IssueApiKeyRequest",
                "type": "object"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Lister les clés d'API du marchand (préfixe seul)",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/merchant/api-keys/{key_id}": {
      "delete": {
        "summary": "Révoquer une clé d'API marchande",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/merchant/webhook": {
      "put": {
        "summary": "Configurer l'URL + le secret des webhooks marchand signés",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "url": {
                    "examples": [
                      "https://shop.example.com/flash/webhook"
                    ],
                    "maxLength": 300,
                    "title": "Url",
                    "type": "string"
                  },
                  "secret": {
                    "maxLength": 128,
                    "minLength": 16,
                    "title": "Secret",
                    "type": "string"
                  }
                },
                "required": [
                  "url",
                  "secret"
                ],
                "title": "ConfigureWebhookRequest",
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Désactiver les webhooks marchand",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/merchant/poster": {
      "get": {
        "summary": "Affiche imprimable (PNG) portant le QR statique du marchand",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/merchant/payments": {
      "get": {
        "summary": "Encaissements du marchand",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/merchant/payments/{payment_id}/refund": {
      "post": {
        "summary": "Rembourser un paiement encaissé (contre-passation)",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 255
            }
          }
        ]
      }
    },
    "/v1/merchant/settlement": {
      "put": {
        "summary": "Configurer le compte bancaire et la fréquence de règlement",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "holder": {
                    "examples": [
                      "SARL Chez Awa"
                    ],
                    "maxLength": 140,
                    "minLength": 1,
                    "title": "Holder",
                    "type": "string"
                  },
                  "iban": {
                    "examples": [
                      "CI93CI0080111301134291200589"
                    ],
                    "maxLength": 40,
                    "minLength": 8,
                    "title": "Iban",
                    "type": "string"
                  },
                  "bank_name": {
                    "examples": [
                      "Ecobank CI"
                    ],
                    "maxLength": 120,
                    "minLength": 1,
                    "title": "Bank Name",
                    "type": "string"
                  },
                  "frequency": {
                    "default": "MANUAL",
                    "examples": [
                      "WEEKLY"
                    ],
                    "title": "Frequency",
                    "type": "string"
                  }
                },
                "required": [
                  "holder",
                  "iban",
                  "bank_name"
                ],
                "title": "ConfigureSettlementRequest",
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/v1/merchant/settlements": {
      "post": {
        "summary": "Déclencher immédiatement un règlement du net accumulé",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Historique des règlements du marchand",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/merchant/settlements/{settlement_id}": {
      "get": {
        "summary": "Relevé détaillé d'un règlement (paiements couverts)",
        "tags": [
          "merchants"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/merchant-payments": {
      "post": {
        "summary": "Payer un marchand (QR statique avec montant, ou QR dynamique)",
        "tags": [
          "merchants"
        ],
        "responses": {
          "201": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "merchant_id": {
                    "maxLength": 64,
                    "minLength": 8,
                    "title": "Merchant Id",
                    "type": "string"
                  },
                  "amount_minor": {
                    "anyOf": [
                      {
                        "exclusiveMinimum": 0,
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "examples": [
                      25000
                    ],
                    "title": "Amount Minor"
                  },
                  "charge_id": {
                    "anyOf": [
                      {
                        "maxLength": 64,
                        "minLength": 8,
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Charge Id"
                  }
                },
                "required": [
                  "merchant_id"
                ],
                "title": "PayMerchantRequest",
                "type": "object"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 255
            }
          }
        ]
      }
    },
...
    "/v1/admin/merchants/{merchant_id}/kyb": {
      "post": {
        "summary": "Back-office : approuver ou rejeter la vérification (KYB) d'un marchand",
        "tags": [
          "admin"
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "4XX": {
            "description": "Erreur métier ou de validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "decision": {
                    "examples": [
                      "approve",
                      "reject"
                    ],
                    "title": "Decision",
                    "type": "string"
                  },
                  "reason": {
                    "default": "",
                    "maxLength": 200,
                    "title": "Reason",
                    "type": "string"
                  }
                },
                "required": [
                  "decision"
                ],
                "title": "ReviewKybRequest",
                "type": "object"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "INSUFFICIENT_FUNDS"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}